minor fixes

+++ src/librexgen/string/simplestring.h
@@ -40,7 +40,7 @@ class SimpleString {
   uchar_t& operator[](const unsigned int& idx) { return characters[idx]; }
 
   const uchar_t& getAt(const unsigned int& idx) const; 
-  int to_binary_string(char* buffer, size_t buffer_size) const;
+  size_t to_binary_string(char* buffer, size_t buffer_size) const;
 
   void newline();
 
@@ -61,7 +61,7 @@ class SimpleString {
   size_t get_buffer_size() const;
 
  private:
-        vector<uchar_t> characters;
+  vector<uchar_t> characters;
 };
 
 #endif /* __cplusplus */
